requirements.txt
appdirs==1.4.3
asgiref==3.2.7
attrs==19.3.0
black==19.10b0
click==7.1.2
Django==3.0.5
django-cors-headers==3.2.1
djangorestframework==3.11.0
entrypoints==0.3
flake8==3.7.9
mccabe==0.6.1
pathspec==0.8.0
pycodestyle==2.5.0
pyflakes==2.1.1
pytz==2020.1
regex==2020.4.4
sqlparse==0.3.1
toml==0.10.0
typed-ast==1.4.1
